[PLAT-18466]: Delete PITR schedule at the start of the Upgrade and recreate after rollback/finalize
Summary: Added changes to delete the existing PITR schedule at the start of any(PG-15/non-PG-15) upgrade and re-create it after upgrade completion(rollback/finalise). As PITR schedules are created as part of xCluster DR operations, I added the same changes for DR's PITR schedule. As a result of this change, customers cannot perform any xcluster DR operation when either cluster is in the upgrade process, except for failover when the target universe is not in the upgrade phase.

Test Plan:
Added unit tests
Tested manually:
 - Create a universe with pitr enabled, perform an upgrade and verify that pitr schedule is deleted from the universe. Perform finalise/rollback and verify a new pitr schedule is created after the task with the same parameters.
 - Created a xcluster DR setup and performed an upgrade on the target universe and verified that the target universe pitr is deleted in the monitoring phase and re-created after the upgrade completion. Also, verified the same for the source universe.

Reviewers: anijhawan, vkumar, hzare, hvontela

Reviewed By: hzare

Subscribers: yugaware

Differential Revision: https://phorge.dev.yugabyte.com/D46581