[PLAT-18403] Add YBA access log filter
Summary:
We have some cases where having an access log would help to troubleshoot customer issues.
One example is - api token is recreated and we need to figure out the client which does that.
Also, it helps to check which exact APIs are slow.

Test Plan:
Start YBA.
Make sure access log is written for all the API calls.

Reviewers: skurapati

Reviewed By: skurapati

Subscribers: yugaware

Differential Revision: https://phorge.dev.yugabyte.com/D46592