[PLAT-18312] Optimize metrics controller
Summary:
Right now metrics controller is one of the hotspots in both CPU sampling and memory allocation sampling profiles.
This happens mostly because we're not using buffered writer and each time label is written we allocate a byte buffer for it.
Also, we're not using gzip compression, hence we have to store uncompressed response in memory.
The diff addresses that.

Test Plan: Make sure metric collection works and metircs endpoint works in both plain text and gzip mode.

Reviewers: vbansal

Reviewed By: vbansal

Subscribers: svc_phabricator, yugaware

Differential Revision: https://phorge.dev.yugabyte.com/D46605