[BACKPORT 2025.1][#26423] CDC: Set retention barriers as a RAFT operation on catalog tablet
Summary:
#####Backport Description
No merge conflicts were encountered.

#####Original Description
D42412 establishes a basic framework to poll catalog tables to figure out changes to the publication / polling list, but does not set the retention barriers on the sys catalog tablet properly. In this revision we introduce changes by which we set the retention barriers on the sys catalog tablet at the time of stream creation as a RAFT operation.

At the time of stream creation, an alter table is performed on all the tables in the stream metadata. This Alter does two things:

  # Sets the `cdc_wal_retention_secs` on all the tablets of the table.
  # Sets retention barriers (WAL, intents, and history) on all the tablets of these tables.

For regular tables, during alter table we send async AlterSchema rpcs (`TabletServiceAdminImpl::AlterSchema`) to the leaders of all the tablets of the table. Here we set the retention barriers on the leader and submit a `CHANGE_METADATA_OP` to apply the same change on other peers of the tablet. During the `CHANGE_METADATA_OP` Apply() path retention barriers are set on the other peers of the tablet. These changes were introduced as a part of D30088.

The same mechanism cannot be employed for catalog tables as AlterSchema is serviced by tablet_service which does not run on master. To set the retention barriers on sys catalog tablet, we directly submit a `CHANGE_METADATA_OP` which will be replicated across all the peers of the sys catalog tablet. The retention barriers will be set on all the peers during the Apply() `CHANGE_METADATA_OP`.  This mechanism is similar to what is followed for regular tablets, the difference being, here we don't explicitly call `SetAllInitialCDCSDKRetentionBarriers()` on the leader and instead let the Apply() path of `CHANGE_METADATA_OP` do this job.

Once this operation is completed, we populate the safe_op_id (the last committed op id at the time of submitting the `CHANGE_METADATA_OP`) in the cdc_state table for sys_catalog_tablet's entry.  This done in the callback of  `CHANGE_METADATA_OP`, hence ensuring that retention barriers have been set on the master leader before populating state table entries.
Jira: DB-15786

Original commit: 5df23574f79a232e41c4bc799075156982485775 / D44843

Test Plan:
./yb_build.sh --cxx-test integration-tests_cdcsdk_ysql-test --gtest_filter CDCSDKYsqlTest.TestPollingPgCatalogTablesAfterLeaderStepDown
./yb_build.sh --cxx-test integration-tests_cdcsdk_consistent_snapshot-test --gtest_filter CDCSDKConsistentSnapshotTest.TestCSStreamFailureRollbackFailureAfterRetentionBarriersWithCatalogTables
./yb_build.sh --cxx-test integration-tests_cdcsdk_ysql-test --gtest_filter CDCSDKYsqlTest.TestFailSettingRetentionBarriersOnApplyForCatalogTable
./yb_build.sh --cxx-test integration-tests_cdcsdk_ysql-test --gtest_filter CDCSDKYsqlTest.TestConcurrentStreamCreationOnCatalogTables

Reviewers: skumar, asrinivasan, stiwary, aagrawal, sergei, #db-approvers

Reviewed By: asrinivasan

Subscribers: ycdcxcluster

Tags: #jenkins-ready

Differential Revision: https://phorge.dev.yugabyte.com/D46615