[PLAT-18503]YBA HA's standby UI does not honor UI restrictions for LDAP user
Summary:
Previous behavior: If there was any error from the /haConfig API, the banner was removed.

Issue identified: On platforms with RBAC enabled, the API returns a 401 Unauthorized error, which is not the expected behavior.

Fix implemented: The logic has been updated to specifically check for a 401 status. If the error is a 401, the error message is now displayed. For all other errors, the banner is removed as before.

Test Plan: Manually verified the changes on both the standby and active instances. Tested with users across different roles ,Super Admin, Admin, and regular user ,to ensure consistent behaviour.

Reviewers: lsangappa

Reviewed By: lsangappa

Differential Revision: https://phorge.dev.yugabyte.com/D46624