[28412] YSQL: Fix PgCatalogVersionTest.NonIncrementingDDLMode with Txn DDL
Summary:
The test `PgCatalogVersionTest.NonIncrementingDDLMode` fails with the following when transactional ddl is enabled.

```
[ts-1] I0828 00:40:22.591221 3948896384 pg_yb_utils.c:3105] pg null=0, nmsgs=3
[ts-1] TRAP: FailedAssertion("false", File: "../../../../../../../src/postgres/src/backend/utils/misc/pg_yb_utils.c", Line: 2667, PID: 33357)
[ts-1] 0   postgres                            0x00000001016bc75c ExceptionalCondition + 280
[ts-1] 1   postgres                            0x00000001016f6b28 YbCatalogModificationAspectsToDdlMode + 140
[ts-1] 2   postgres                            0x00000001016f9f1c YbGetDdlMode + 3712
[ts-1] 3   postgres                            0x0000000101709454 YBTxnDdlProcessUtility + 112
.....
```

An even simpler repro is:
```
yugabyte=# CREATE TABLE t1(a int);
CREATE TABLE
yugabyte=# SET yb_make_next_ddl_statement_nonincrementing TO TRUE;
SET
yugabyte=# CREATE INDEX idx3 ON t1(a);
server closed the connection unexpectedly
	This probably means the server terminated abnormally
	before or while processing the request.
The connection to the server was lost. Attempting reset: Failed.
The connection to the server was lost. Attempting reset: Failed.
```

The trap that is getting hit is this
```
YbCatalogModificationAspectsToDdlMode(uint64_t catalog_modification_aspects)
{
	YbDdlMode	mode = catalog_modification_aspects;

	switch (mode)
	{
		case YB_DDL_MODE_NO_ALTERING:
			yb_switch_fallthrough();
		case YB_DDL_MODE_SILENT_ALTERING:
			yb_switch_fallthrough();
		case YB_DDL_MODE_VERSION_INCREMENT:
			yb_switch_fallthrough();
		case YB_DDL_MODE_BREAKING_CHANGE:
                        yb_switch_fallthrough();
                case YB_DDL_MODE_AUTONOMOUS_TRANSACTION_CHANGE_VERSION_INCREMENT:
			return mode;
	}
	Assert(false); <-------------------------------------------------------------- Causes the crash.
	return YB_DDL_MODE_BREAKING_CHANGE;
}
```

#### Cause
As part of D43199 / 8eb1e718f864, we updated the logic to always run `CREATE INDEX` in autonomous transactions if it isn't running in an explicit transaction block. For more info, see case 2a and 2b of the summary in the linked commit / revision. To implement that, we introduced the `YB_SYS_CAT_MOD_ASPECT_AUTONOMOUS_TRANSACTION_CHANGE` as the indication that the DDL must be executed as an autonomous transaction.

The bug is that the change in D43199 / 8eb1e718f864 didn't factor in other `YbSysCatalogModificationAspect` values and assumed that autonomous transactions are only required for CREATE INDEX which is **always** a breaking change. This assumption is wrong and breaks when the user explicitly makes the DDL non-breaking or non-incrementing.

#### Fix - Approach 1
Define two variants of YbDdlMode for each option:

```
	YB_DDL_MODE_NO_ALTERING = 0,

	YB_DDL_MODE_AUTONOMOUS_TRANSACTION_NO_ALTERING = YB_SYS_CAT_MOD_ASPECT_AUTONOMOUS_TRANSACTION_CHANGE,

	YB_DDL_MODE_SILENT_ALTERING = YB_SYS_CAT_MOD_ASPECT_ALTERING_EXISTING_DATA,

	YB_DDL_MODE_AUTONOMOUS_TRANSACTION_SILENT_ALTERING = YB_SYS_CAT_MOD_ASPECT_ALTERING_EXISTING_DATA | YB_SYS_CAT_MOD_ASPECT_AUTONOMOUS_TRANSACTION_CHANGE,

	YB_DDL_MODE_VERSION_INCREMENT = (YB_SYS_CAT_MOD_ASPECT_ALTERING_EXISTING_DATA |
									 YB_SYS_CAT_MOD_ASPECT_VERSION_INCREMENT),

	YB_DDL_MODE_AUTONOMOUS_TRANSACTION_CHANGE_VERSION_INCREMENT =
		(YB_SYS_CAT_MOD_ASPECT_ALTERING_EXISTING_DATA |
		 YB_SYS_CAT_MOD_ASPECT_VERSION_INCREMENT |
		 YB_SYS_CAT_MOD_ASPECT_AUTONOMOUS_TRANSACTION_CHANGE),

        YB_DDL_MODE_BREAKING_CHANGE = (YB_SYS_CAT_MOD_ASPECT_ALTERING_EXISTING_DATA |
								   YB_SYS_CAT_MOD_ASPECT_VERSION_INCREMENT |
								   YB_SYS_CAT_MOD_ASPECT_BREAKING_CHANGE),

         YB_DDL_MODE_AUTONOMOUS_TRANSACTION_BREAKING_CHANGE = (YB_SYS_CAT_MOD_ASPECT_ALTERING_EXISTING_DATA |
								   YB_SYS_CAT_MOD_ASPECT_VERSION_INCREMENT |
								   YB_SYS_CAT_MOD_ASPECT_BREAKING_CHANGE |
                                                                   YB_SYS_CAT_MOD_ASPECT_AUTONOMOUS_TRANSACTION_CHANGE),
```

This way we'll be able to represent all possible combinations for create index depending on the values of GUC `yb_make_next_ddl_statement_nonincrementing` and `yb_make_next_ddl_statement_nonbreaking`.

#### Fix - Approach 2 (Selected)
Change the approach of returning the requirement of autonomous transactions via DDL mode and instead just accept a boolean pointer in `YbGetDdlMode`.

Approach 2 has been implemented here due to the simplicity.

Test Plan:
Enabled transactional DDL and ran the test:

```
./yb_build.sh --cxx-test pgwrapper_pg_catalog_version-test --gtest_filter PgCatalogVersionTest.NonIncrementingDDLMode
```

Also tried the manual repro steps shared in summary.

Reviewers: myang

Reviewed By: myang

Subscribers: yql

Differential Revision: https://phorge.dev.yugabyte.com/D46668