[#28000] YSQL: Make COMMIT a DDL during YSQL upgrade in all cases
Summary:
As part of D43861 / 6af351866073, we made `COMMIT` a DDL when running during YSQL upgrade. However, this was not done when transactional DDL is enabled. It was expected that with transactional DDL, we will always increment the catalog version as a precaution even if it didn't execute any DDL. Incrementing the catalog version is required to capture any invalidation messages.

The long term plan to always increment the catalog version when a transaction block is created is left for the future. For now, this revision makes the COMMIT a DDL during ysql upgrade even when Txn DDL is enabled.

This change helps fix `PgCatalogVersionTest.InvalMessageYsqlUpgradeCommit1` and similar tests when transactional ddl is enabled.
Jira: DB-17619

Test Plan:
Enabled transactional ddl and ran the following tests:

```
./yb_build.sh --cxx-test pgwrapper_pg_catalog_version-test --gtest_filter PgCatalogVersionTest.InvalMessageYsqlUpgradeCommit1
./yb_build.sh --cxx-test pgwrapper_pg_catalog_version-test --gtest_filter PgCatalogVersionTest.InvalMessageYsqlUpgradeCommit2
./yb_build.sh --cxx-test pgwrapper_pg_catalog_version-test --gtest_filter PgCatalogVersionTest.InvalMessageYsqlUpgradeCommit3
./yb_build.sh --cxx-test pgwrapper_pg_catalog_version-test --gtest_filter PgCatalogVersionTest.InvalMessageYsqlUpgradeCommit4
```

Reviewers: myang

Reviewed By: myang

Subscribers: yql

Differential Revision: https://phorge.dev.yugabyte.com/D46693