[BACKPORT 2025.1][#28269] docdb: Pass tablespace-tagged status tablets in GetTransactionStatusTablets response
Summary:
In preparation for adding categorization of transactions into a new tablespace-local localities,
client code needs transaction status tablets grouped by tablespace, instead of just an unlabeled list of
region-local status tablet ids that we send today:
```
message GetTransactionStatusTabletsResponsePB {
  optional MasterErrorPB error = 1;

  repeated bytes global_tablet_id = 2;
  repeated bytes placement_local_tablet_id = 3;
}
```

This change adds a `placement_local_tablespaces` field to `GetTransactionStatusTabletsResponsePB`
containing a mapping of tablespace oid to placement and list of transaction status tablet ids. This is
currently only used under a test gflag, but will be used in future changes to add tablespace-local locality.

The old list of tablets was also renamed from placement_local_... to region_local_... to better
reflect the contents and distinguish it from the new list.

**Upgrade/Downgrade Safety**
The new `placement_local_tablespaces` is only sent if the newly added
`TEST_enable_tablespace_based_transaction_placement` is enabled. This will be changed to
an autoflag in the future, so that the tablespace-local locality feature to be added, which depends
on `placement_local_tablespaces` being set, will also be guarded by the same autoflag.

To ensure that a GetTransactionStatusTablets response with tablespace information is actually sent
post-upgrade, even if tserver is upgraded before master, the transaction table version is bumped
when the flag is changed.
Jira: DB-17944

Original commit: 934088c16ce40844f1629ad0408d5d7b2f4a7044 / D45963

Test Plan:
Added test:
```
./yb_build.sh --cxx-test pgwrapper_geo_transactions-test --gtest_filter 'GeoTransactionsTablespaceBasedSelectionCandidatesTest.TestCandidates'
```

Reviewers: sergei, rthallam

Reviewed By: rthallam

Subscribers: ybase, yql

Differential Revision: https://phorge.dev.yugabyte.com/D46733