[#28573] Docdb: ReleaseObjectLocks should not persist db_catalog_invalidation_messages
Summary:
Master persists the ReleaseObjectLockRequestPB to the sys-catalog to ensure that
a release operation can be resumed in the event of a master failover. However,
including the db_catalog invalidation messages (which is piggy-backed onto
the ReleaseObjectLocks Rpc) makes us log too much data to the WALs.

This may cause the master to replay a lot of WALs after a restart, and cause
unavailability of the master tablet.

The change is to ensure that we Populate the DB catalog information into the
release request only when the request is being sent to the Tservers, and not
when we are (initially) persisting the ReleaseObjectLockRequestPB.

**Upgrade/Downgrade Safety**
Should be ok to downgrade/upgrade if the table locks is not enabled during the ugprade/downgrade (which is default as of now).

If we downgrade while the table locks feature is enabled, it is possible that if a new-master persists a release request which is later completed by an older-master, then the reqeusts forwarded to the TServer may not have the db_catalog_invalidation message as part of the ReleaseRequest.
Jira: DB-18267

Test Plan:
yb_build.sh fastdebug --cxx-test object_lock-test
yb_build.sh fastdebug --cxx-test pb_object_locks-test

Verify the fix using yb-stress-test

Reviewers: rthallam, bkolagani, zdrudi

Reviewed By: zdrudi

Subscribers: ybase

Differential Revision: https://phorge.dev.yugabyte.com/D46777