[#28000] YSQL: Fix tests with transactional ddl - part 1
Summary:
Transactional DDL introduces the ability to run DDL as part of the transaction block similar to DMLs. Before the feature, DDL used to run in
autonomous transactions and commit independently from the overall transaction block it was part of.

As part of enabling txn ddl by default, we first need to fix any failing unit tests. This revision fixes the following tests:
1. TestPgRegressCursor#testPgRegressCursor
2. TestPgRegressPgTypesUDT
3. PgStatActivityTest.DDLInsideDMLTransaction

For the regress tests, we are enabling transactional DDL by adding the GFlag. They'll be removed once we enable it by default.
Jira: DB-17619

Test Plan:
```
./yb_build.sh --java-test 'org.yb.pgsql.TestPgRegressPgTypesUDT#schedule'
./yb_build.sh --java-test 'org.yb.pgsql.TestPgRegressCursor#testPgRegressCursor'
./yb_build.sh --cxx-test pgwrapper_pg_stat_activity-test --gtest_filter PgStatActivityTest.DDLInsideDMLTransaction
```

Reviewers: kramanathan, myang, #db-approvers

Reviewed By: myang, #db-approvers

Subscribers: svc_phabricator, jason, yql

Differential Revision: https://phorge.dev.yugabyte.com/D46698