[#28000] YSQL: Enable txn ddl in TestPgRegressInsertOnConflict
Summary:
Transactional DDL introduces the ability to run DDL as part of the transaction block similar to DMLs. Before the feature, DDL used to run in autonomous transactions and commit independently from the overall transaction block it was part of. This also means that upon `ABORT`, the schema changes made by the DDLs are also rolled back which wasn't the case earlier.

This revision updates the TestPgRegressInsertOnConflict test to pass when transactional DDL is enabled. This involves updating both the SQL script and the expected output. Transactional DDL flag has been set in the test to enable the feature. It'll be removed once the feature is enabled by default (marked as TODO).
Jira: DB-17619

Test Plan:
Jenkins: test regex: .*TestPgRegressInsertOnConflict.*

./yb_build.sh --java-test 'org.yb.pgsql.TestPgRegressInsertOnConflict'
./yb_build.sh --java-test 'org.yb.pgsql.TestPgRegressInsertOnConflictBatch0'
./yb_build.sh --java-test 'org.yb.pgsql.TestPgRegressInsertOnConflictBatch1'
./yb_build.sh --java-test 'org.yb.pgsql.TestPgRegressInsertOnConflictBatch2'
./yb_build.sh --java-test 'org.yb.pgsql.TestPgRegressInsertOnConflictBatch3'
./yb_build.sh --java-test 'org.yb.pgsql.TestPgRegressInsertOnConflictBatch9'

Reviewers: jason

Reviewed By: jason

Subscribers: yql

Differential Revision: https://phorge.dev.yugabyte.com/D46736