[#28000] YSQL: Disable txn ddl in tests which use savepoints with DDLs
Summary:
Transactional DDL introduces the ability to run DDL as part of the transaction block similar to DMLs. Before the feature, DDL used to run in autonomous transactions and commit independently from the overall transaction block it was part of. This also means that upon ABORT, the schema changes made by the DDLs are also rolled back which wasn't the case earlier. At the moment, we do not support savepoints for DDLs. As a result, when transactional DDL is enabled, we throw an error when a savepoint is attempted after running a DDL in a transaction block

This revision explicitly disables transactional DDL in the tests which use savepoint with DDLs. This is to ensure no test fails when we enable transactional DDL by default. TODOs have been left to re-enable txn ddl once savepoint support is added.
1. TestPgRegressPgTransactions
2. TestPgRegressMatview
3. TestPgRegressParallel
4. TestPgRegressPgMiscIndependent
Jira: DB-17619

Test Plan:
./yb_build.sh --java-test 'org.yb.pgsql.TestPgRegressMatview#testPgRegressMatview'
./yb_build.sh --java-test 'org.yb.pgsql.TestPgRegressParallel#testPgRegressParallel'
./yb_build.sh --java-test 'org.yb.pgsql.TestPgRegressPgMiscIndependent#testPgRegressPgMiscIndependent'
./yb_build.sh --java-test 'org.yb.pgsql.TestPgRegressPgTransactions'

Reviewers: jason, myang

Reviewed By: myang

Subscribers: yql

Differential Revision: https://phorge.dev.yugabyte.com/D46811