[PLAT-18572] Stopped onprem node is removed from blacklist after release node
Summary:
The hardening diff https://phorge.dev.yugabyte.com/D46312 replaced fetch-by-node UUID with fetch-by-name. As the node is no longer associated with the universe when this subtask is run, fetch-by-name does not return
the node and is not not removed from the nodes to be un-blacklisted.

Test Plan: Itest must pass.

Reviewers: yshchetinin, anijhawan, hzare

Reviewed By: anijhawan, hzare

Subscribers: hzare, yugaware

Differential Revision: https://phorge.dev.yugabyte.com/D46824