[#28000] YSQL: Fix test TestPgRegressMisc#testPgRegressMiscIndependent with txn ddl
Summary:
Transactional DDL introduces the ability to run DDL as part of the transaction block similar to DMLs. Before the feature, DDL used to run in
autonomous transactions and commit independently from the overall transaction block it was part of.

The `ANALYZE` statement when executed outside of an explicit transaction block can be executed in multiple YB transactions. Those cases are:
1. Outside of txn block and processing more than 1 relation
2. Autovacuum worker (Unsupported in YSQL)

1 PG (& YB) txn is created per relation so that we can release the locks for that relation after we are done processing it instead of waiting for all
relations to be processed.

Without transactional DDL, since the DDLs were only committed at the end of the statement, only one DDL txn was created for the ANALYZE command
regardless of the number of relations involved. As a result, the ANALYZE command would always increment the catalog version by 1.

With transactional DDL, there is no concept of DDL txn. So the ANALYZE of each relation leads to a catalog version increment. This revision updates
the expected catalog version bump after every ANALYZE statement in the TestPgRegressMisc#testPgRegressMiscIndependent regress test.
Jira: DB-17619

Test Plan:
Jenkins: test regex: .*TestPgRegressMisc.*

./yb_build.sh --java-test 'org.yb.pgsql.TestPgRegressMisc#testPgRegressMiscIndependent'

Reviewers: myang

Reviewed By: myang

Subscribers: yql

Differential Revision: https://phorge.dev.yugabyte.com/D46891