[#28086] CDC: Reduce log spam due to tablet not found warning
Summary:
The log spam happens when a (stream_id, tablet_id) entry is no longer present in cdc stable table (or is expired) however, it is present in cache 'tablet_checkpoints'. The RemoveXReplTabletMetrics() then tries to clean up the metric metadata for such stream for the concerned tablet. However there can be the case that there still exist a mapping from tablet_id to tablet_peer ptr (in tablet server's TSTabletManager) but underlying tablet handle is not present. Thus when RemoveXReplTabletMetrics() gets a null tablet ptr using non-null tablet_peer ptr, it logs the concerned warning.

Given, that this whole logic is handled as part of UpdatePeersAndMetrics() which runs periodically and 'tablet_checkpoints' cache doesn't clear such entry, the same logic keeps hitting the warning leading to log spam.

This problem anyways goes away if the tablet server gets restarted (as the in-mem mapping from tablet_id to tablet_peer ptr gets cleared out).
So as a quick workaround, this diff changes warning to log every 300 sec or log everytime if verbose logging is set to level 4 or above. This helps in reducing the log spam.

Jira: DB-17715

Test Plan: Jenkins: test regex: CDC*

Reviewers: asrinivasan, sumukh.phalgaonkar, skumar, stiwary

Reviewed By: sumukh.phalgaonkar

Subscribers: ycdcxcluster

Tags: #jenkins-ready

Differential Revision: https://phorge.dev.yugabyte.com/D46768