[#28000] YSQL: Fix test TestPgRegressThirdPartyExtensionsPgHintPlan with txn ddl
Summary:
Transactional DDL introduces the ability to run DDL as part of the transaction block similar to DMLs. Before the feature, DDL used to run in
autonomous transactions and commit independently from the overall transaction block it was part of.

This revision enables transactional DDL in the TestPgRegressThirdPartyExtensionsPgHintPlan test and fixes it to work with txn ddl. When txn ddl is
enabled, the function `YBIncrementDdlNestingLevel` is no longer applicable as it is only used in the old model of separate ddl txn.
Jira: DB-17619

Test Plan: ./yb_build.sh --java-test 'org.yb.pgsql.TestPgRegressThirdPartyExtensionsPgHintPlan#schedule'

Reviewers: myang, kfranz

Reviewed By: myang

Subscribers: yql

Differential Revision: https://phorge.dev.yugabyte.com/D46927