[BACKPORT 2025.1.1][#27177] CDC: Workaround for logical replication during PG major version upgrade
Summary:
##### Backport Description
Encountered merge conflicts in `guc.c` due to some other gucs not being backported to 2025.1

##### Original Description
During PG 15 upgrade, the catalog tables are deleted and re-created. Logical replication relies on reading the catalog tables as of past time when the stream is lagging. Walsender process reads catalog tables as of past time in 3 places:

  #  At the startup it reads the list of tables in the publication as of `last_pub_refresh_time`.
  # When walsender receives a pub refresh record from the virtual WAL it reads the list of tables in the publication as of the the `pub_refresh_time`.
  # When walsender receives a DDL record from the virtual WAL, it reads the schema of the table as of the commit time of next received DML.

If the stream is lagging to an extent, such that after the PG 15 upgrade, a catalog read is performed by the walsender for any of the three reasons mentioned above,  with a read time lying in the PG 11 era, the following error is encountered and the slot is rendered useless:

```
ERROR:  catalog version for database 16640 was not found.
HINT:  Database might have been dropped by another user
STATEMENT:  START_REPLICATION SLOT "test_slot" LOGICAL 0/2D3B0 ("proto_version" '1', "publication_names" 'test_pub', "messages" 'true')
```

To provide a workaround for this problem, this diff introduces a GUC `yb_ignore_read_time_in_walsender`. When this GUC is set, if an error is encountered for the above mentioned first two reasons, then we reset the read time and fetch the list of tables in the publication as of the current time.

This GUC should be set carefully and only till the time the process of upgrading logical replication streams is complete (i.e till the `yb_restart_time` of all the streams crosses the time of upgrade completion). Moreover this GUC should be set only after ensuring that no more DDLs (including `ALTER PUBLICATION`) will be encountered by the walsender; i.e. stop the DDLs at time 'T' and set this GUC only after the restart time crosses 'T'. This way we will not hit the mentioned error due to the third reason mentioned above.

The connector PR to set this GUC in the walsender session based on a config: https://github.com/yugabyte/debezium/pull/185
Jira: DB-16663

Original commit: c2e575feb09a5c2ca9d8bde042cd3b5bb7e2ab42 / D46652

Test Plan:
Jenkins: urgent
Manual Testing. The test was as follows:
  # Create a yugabyted single node universe on 2024.2
  # Create a table, replication slot and deploy the connector.
  # Consume the changes from the table for some time and then delete the connector. Perform some inserts to accumulate lag.
  # Compile the changes in this revision in a separate repo.
  # Upgrade the yugabyted from 2024.2 to compiled binaries (PG 15) following these [[ https://docs.yugabyte.com/stable/reference/configuration/yugabyted/#upgrade-a-yugabytedb-universe | steps ]]
  # Re-deploy the connector using the same slot on the upgraded yugabyted. The connector will set the GUC. We should be able to consume the changes both from before and after the upgrade.

Reviewers: asrinivasan, skumar, stiwary, aagrawal

Reviewed By: skumar

Subscribers: jason, yql, ycdcxcluster

Tags: #jenkins-ready

Differential Revision: https://phorge.dev.yugabyte.com/D46949