[PLAT-18589] Add tooltip and helper text to dynatrace telemetry provider form
Summary:
Simple UI change which adds a tooltip, placeholder text and helper text to the dynatrace telemetry
provider form.
This change is made to improve the user experience when creating the telemetry providers.

Also includes a minor change for showing the supported export options (metrics/logs) next to each telemetry provider in the radio group.

Test Plan:
Create a dynatrace telemetry provider.
Verify the UI elements are shown as expected.
{F396078}
{F396080}
{F396081}
{F396601}

Reviewers: kkannan, lsangappa, skurapati

Reviewed By: skurapati

Differential Revision: https://phorge.dev.yugabyte.com/D46915