[#28537] YSQL: Make sure that TimeZone is not lowercased in connection manager
Summary:
We lowercase all the GUC variable names in connection manager, which results in the
ParameterStatus packets also having the lowercased name. Among the variables that pgJDBC does exact
string matching for, the only one where case matters is TimeZone. Until we are able to store the
variables in the correct case in connection manager, this is a workaround to ensure that TimeZone is
interpreted correctly by pgJDBC
Jira: DB-18236

Test Plan:
Run the following test added in this diff:

```
./yb_build.sh release --enable-ysql-conn-mgr-test --java-test org.yb.ysqlconnmgr.TestSessionParameters#testTimeZoneSetting
```

Jenkins: urgent, test regex: .*ysqlconnmgr.*

Reviewers: skumar, mkumar, asrinivasan, vikram.damle

Reviewed By: skumar

Subscribers: svc_phabricator, yql

Differential Revision: https://phorge.dev.yugabyte.com/D46979